Fix frame extending past viewport when zoomed in
When a user zooms in (`ctrl/cmd` + `+`), the Vomnibar frame will extend
down past the viewport. The height is set to calc(100% - 70px) on the
actual iframe, and we can't reach it with this CSS. Setting a max-height
that is effectively the same, but measured in viewport units, on the
Vomnibar container keeps it from going past the bottom of the viewport.